requirements.txt
flask
pillow 
exifread face_recognition 
scikit-learn